Move shadertools files into submodule structure
