+++ src/core/model.js
@@ -7,8 +7,8 @@ import {isWebGL} from '../webgl-utils';
 import {getUniformsTable, areUniformsEqual} from '../webgl/uniforms';
 import {getDrawMode} from '../geometry/geometry';
 import Object3D from '../core/object-3d';
-import {MODULAR_SHADERS} from '../shadertools/shaders';
-import {assembleShaders} from '../shadertools';
+import {MODULAR_SHADERS} from '../shadertools/src/shaders';
+import {assembleShaders} from '../shadertools/src';
 import {addModel, removeModel, logModel, getOverrides} from '../debug/seer-integration';
 import {log, formatValue, isObjectEmpty} from '../utils';
 import assert from '../utils/assert';
