+++ src/shadertools/src/lib/assemble-shaders.js
@@ -1,8 +1,8 @@
 import {resolveModules, getShaderModule} from './shader-modules';
 import {getPlatformShaderDefines, getVersionDefines} from './platform-defines';
 import {MODULE_INJECTORS_VS, MODULE_INJECTORS_FS} from '../modules/module-injectors';
-import assert from '../../utils/assert';
-import {log} from '../../utils';
+import assert from '../../../utils/assert';
+import {log} from '../../../utils';
 
 const VERTEX_SHADER = 'vs';
 const FRAGMENT_SHADER = 'fs';
