+++ src/shadertools/src/lib/platform-defines.js
@@ -1,5 +1,7 @@
-import {getContextInfo} from '../../webgl';
-import {hasFeature, canCompileGLGSExtension, FEATURES} from '../../webgl-context/context-features';
+import {getContextInfo} from '../../../webgl';
+import {
+  hasFeature, canCompileGLGSExtension, FEATURES
+} from '../../../webgl-context/context-features';
 
 export function checkRendererVendor(debugInfo, gpuVendor) {
   const {vendor, renderer} = debugInfo;
