+++ src/shadertools/src/lib/shader-modules.js
@@ -1,4 +1,4 @@
-import assert from '../../utils/assert';
+import assert from '../../../utils/assert';
 
 const shaderModules = {};
 let defaultShaderModules = [];
