+++ src/shadertools/src/modules/diffuse/diffuse.js
