+++ src/shadertools/src/modules/lighting/README.md
