+++ src/shadertools/src/modules/lighting/lighting-common.glsl.js
