+++ src/shadertools/src/modules/lighting/lighting-fragment.glsl.js
