+++ src/shadertools/src/modules/lighting/lighting-vertex.glsl.js
