+++ src/shadertools/src/modules/lighting/lighting.js
