+++ src/shadertools/src/modules/module-injectors.js
