+++ src/shadertools/src/modules/picking/picking.js
