+++ src/shadertools/src/shaders/default-fragment.glsl.js
