+++ src/shadertools/src/shaders/default-uniforms.js
